[SPARK-34087][3.1][SQL] Fix memory leak of ExecutionListenerBus
This PR proposes an alternative way to fix the memory leak of `ExecutionListenerBus`, which would automatically clean them up.

Basically, the idea is to add `registerSparkListenerForCleanup` to `ContextCleaner`, so we can remove the `ExecutionListenerBus` from `LiveListenerBus` when the `SparkSession` is GC'ed.

On the other hand, to make the `SparkSession` GC-able, we need to get rid of the reference of `SparkSession` in `ExecutionListenerBus`. Therefore, we introduced the `sessionUUID`, which is a unique identifier for SparkSession, to replace the  `SparkSession` object.

Note that, the proposal wouldn't take effect when `spark.cleaner.referenceTracking=false` since it depends on `ContextCleaner`.

Fix the memory leak caused by `ExecutionListenerBus` mentioned in SPARK-34087.

Yes, save memory for users.

Added unit test.

Closes #31839 from Ngone51/fix-mem-leak-of-ExecutionListenerBus.

Authored-by: yi.wu <yi.wudatabricks.com>
Signed-off-by: HyukjinKwon <gurwls223apache.org>

Closes #31881 from Ngone51/SPARK-34087-3.1.

Authored-by: yi.wu <yi.wu@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 5f46bee4d09a655e38a89f9134138f1826855870)
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>